[tests] Moved auto-group tests to base tests
